// types/enums.ts
export enum Images {
  Placeholder = "/images/placeholder.svg",
}

export enum Services {
  Hotel = 'hotel',
  Flights = 'flights',
  BookTransfer = 'book-transfer',
  TourPackage = 'tour-package',
  Skyscanner = 'skyscanner'
}